Create and check in a MANIFEST.
It is a generated file, but it is supposed to be generated the same
everywhere, so it should be in version control.  Now we can see it
as it changes and its a lot easier to see if something incorrect
sneaks in.

It also stops Build.PL from barking about it.